Feat/profile picture

+++ src/accounts/accountsController.js
@@ -24,7 +24,9 @@ async function uploadProfilePicture(req, res, next) {
 
     new responseHandler(res, account, 200, RESPONSE_MESSAGE.SUCCESS);
   } catch (error) {
-    deleteFile(req.file.path);
+    if (req.file) {
+      deleteFile(req.file.path);
+    }
     return next(error);
   }
 }
